refactor: split large components into subcomponents
