+++ web/src/components/node_menu/NamespaceList.tsx
@@ -15,8 +15,10 @@ import {
 import NodeInfo from "./NodeInfo";
 import { isEqual } from "lodash";
 import useMetadataStore from "../../stores/MetadataStore";
-import { KeyboardArrowLeft, AddCircleOutline } from "@mui/icons-material";
-import { usePanelStore } from "../../stores/PanelStore";
+import { KeyboardArrowLeft } from "@mui/icons-material";
+
+import NoSelectionContent from "./subcomponents/NoSelectionContent";
+import InfoBox from "./subcomponents/InfoBox";
 
 type NamespaceTree = {
   [key: string]: {
@@ -404,206 +406,6 @@ const namespaceStyles = (theme: any) =>
     }
   });
 
-const NoSelectionContent = memo(function NoSelectionContent({
-  searchTerm,
-  selectedPathString,
-  minSearchTermLength
-}: {
-  searchTerm: string;
-  selectedPathString: string;
-  minSearchTermLength: number;
-}) {
-  const handleViewChange = usePanelStore((state) => state.handleViewChange);
-
-  const openPacksPanel = useCallback(() => {
-    handleViewChange("packs");
-  }, [handleViewChange]);
-
-  return (
-    <div className="no-selection">
-      {searchTerm.length > minSearchTermLength ? (
-        <>
-          <p>
-            {selectedPathString ? (
-              <>
-                Nothing found in this namespace for
-                <strong className="highlighted-text">
-                  {" "}
-                  &quot;{searchTerm}&quot;
-                </strong>
-              </>
-            ) : (
-              <>
-                Nothing found for
-                <strong className="highlighted-text">
-                  {" "}
-                  &quot;{searchTerm}&quot;
-                </strong>
-              </>
-            )}
-          </p>
-          <ul className="no-results">
-            {selectedPathString && (
-              <li>
-                click on <span className="highlighted">highlighted </span>
-                namespaces to find results.
-              </li>
-            )}
-            <li>just start typing to enter a new search term</li>
-            <li>clear search by clicking the clear button</li>
-          </ul>
-        </>
-      ) : (
-        <div className="explanation">
-          <h4>Node Search</h4>
-          <ul>
-            <li>
-              Browse through available nodes by selecting namespaces from the
-              menu on the left
-            </li>
-            <li>
-              Add nodes to your workflow by:
-              <ul>
-                <li>Clicking on the desired node</li>
-                <li>Or dragging it directly onto the canvas</li>
-              </ul>
-            </li>
-          </ul>
-        </div>
-      )}
-      <div className="node-packs-info">
-        <Button
-          startIcon={<AddCircleOutline />}
-          size="small"
-          variant="outlined"
-          onClick={openPacksPanel}
-          style={{
-            marginTop: "0.5em",
-            marginBottom: "0.5em",
-            textTransform: "none",
-            lineHeight: "1.5",
-            padding: "15px"
-          }}
-        >
-          Install additional node packs
-        </Button>
-      </div>
-
-      <Typography variant="h6">Let us know what&apos;s missing</Typography>
-      <ul className="no-results">
-        <li>
-          Join our{" "}
-          <a
-            href="#"
-            onClick={(e) => {
-              e.preventDefault();
-              window.open(
-                "https://discord.gg/WmQTWZRcYE",
-                "_blank",
-                "noopener,noreferrer"
-              );
-            }}
-          >
-            Discord
-          </a>
-        </li>
-        <li>
-          Join the{" "}
-          <a
-            href="#"
-            onClick={(e) => {
-              e.preventDefault();
-              window.open(
-                "https://forum.nodetool.ai",
-                "_blank",
-                "noopener,noreferrer"
-              );
-            }}
-          >
-            Nodetool Forum
-          </a>
-        </li>
-      </ul>
-    </div>
-  );
-});
-
-const InfoBox = memo(function InfoBox({
-  searchTerm,
-  minSearchTermLength,
-  selectedPathString,
-  searchResults,
-  allSearchMatches,
-  metadata,
-  totalNodes
-}: {
-  searchTerm: string;
-  minSearchTermLength: number;
-  selectedPathString: string;
-  searchResults: NodeMetadata[];
-  allSearchMatches: NodeMetadata[];
-  metadata: NodeMetadata[];
-  totalNodes: number;
-}) {
-  return (
-    <Box className="info-box">
-      <Tooltip
-        title={
-          <div style={{ color: "#eee", fontSize: "1.25em" }}>
-            {selectedPathString && (
-              <div>Current namespace: {searchResults?.length} nodes</div>
-            )}
-            {searchTerm.length > minSearchTermLength ? (
-              <>
-                <div>Total search matches: {allSearchMatches.length}</div>
-                <div
-                  style={{
-                    fontSize: "0.8em",
-                    color: "#aaa",
-                    marginTop: "0.5em"
-                  }}
-                ></div>
-              </>
-            ) : (
-              <>
-                <div>Total available: {totalNodes} nodes</div>
-                <div
-                  style={{
-                    fontSize: "0.8em",
-                    color: "#aaa",
-                    marginTop: "0.5em"
-                  }}
-                ></div>
-              </>
-            )}
-          </div>
-        }
-        enterDelay={TOOLTIP_ENTER_DELAY}
-        leaveDelay={TOOLTIP_LEAVE_DELAY}
-        enterNextDelay={TOOLTIP_ENTER_NEXT_DELAY}
-        placement="bottom"
-      >
-        <Typography className="result-info">
-          {searchTerm.length > minSearchTermLength ? (
-            <>
-              <span>{searchResults.length}</span> /{" "}
-              <span>
-                {searchTerm.length > minSearchTermLength
-                  ? allSearchMatches.length
-                  : metadata.length}
-              </span>
-            </>
-          ) : (
-            <span>
-              {selectedPathString ? searchResults.length : totalNodes}
-            </span>
-          )}
-          <span className="result-label">nodes</span>
-        </Typography>
-      </Tooltip>
-    </Box>
-  );
-});
 
 const NamespaceList: React.FC<NamespaceListProps> = ({
   namespaceTree,
